Add __dirname before model path in README
I got a mysterious error `TypeError: Class constructor Model cannot be invoked without 'new'` and tried to debug for a while until I realized the reason of the error. My models are in files like `'./**/model.js'` so I adapted my case as described in the readme... but actually a file `./node_modules/sequelize/lib/model.js` match this regex unless the above regex is prefixed by `__dirname`. This is why I think it would be great to add the __dirname.